refactor: types 파일들 export 방식 통일
