+++ client/src/types/TransactionForm.ts
@@ -1,4 +1,4 @@
-import { Options } from './options';
+import { Options } from './dropdownOptions';
 
 export interface ITransactionForm {
   classify: boolean;
