+++ client/src/types/account.ts
@@ -1,5 +1,7 @@
-export default interface Account {
+export interface Account {
   id: number;
   name: string;
   color: string;
 }
+
+export default Account;
