+++ client/src/types/buttonTypes.ts
@@ -2,3 +2,5 @@ export interface ModalCommonButton {
   onClick?: () => void;
   children?: React.ReactNode;
 }
+
+export default ModalCommonButton;
