+++ client/src/types/category.ts
@@ -1,5 +1,7 @@
-export default interface Category {
+export interface Category {
   id: number;
   name: string;
   color: string;
 }
+
+export default Category;
