+++ client/src/types/income.ts
@@ -2,7 +2,7 @@ import Expenditure from './expenditure';
 import Category from './category';
 import Account from './account';
 
-export default interface Income {
+export interface Income {
   id: number;
   amount: number;
   content: string;
@@ -15,3 +15,5 @@ export default interface Income {
 export const isIncome = (transaction: Income | Expenditure): transaction is Income => {
   return (transaction as Income).content !== undefined;
 };
+
+export default Income;
