+++ client/src/types/lineChartValue.ts
@@ -2,3 +2,5 @@ export interface ITransaction {
   date: Date;
   value: number;
 }
+
+export default ITransaction;
