+++ client/src/types/query.ts
@@ -1,7 +1,9 @@
-export default interface Query {
+export interface Query {
   startDate?: string | string[] | null;
   endDate?: string | string[] | null;
   incomeCategory?: string | string[] | null;
   expenditureCategory?: string | string[] | null;
   account?: string | string[] | null;
 }
+
+export default Query;
